+++ curve25519-dalek/src/backend/vector/avx2.rs
@@ -9,7 +9,7 @@
 // - isis agora lovecruft <isis@patternsinthevoid.net>
 // - Henry de Valence <hdevalence@hdevalence.ca>
 
-#![doc = include_str!("../../../../docs/avx2-notes.md")]
+#![doc = include_str!("../../../docs/avx2-notes.md")]
 
 pub(crate) mod field;