+++ ed25519-dalek/benches/ed25519_benchmarks.rs
@@ -7,7 +7,7 @@
 // Authors:
 // - isis agora lovecruft <isis@patternsinthevoid.net>
 
-use criterion::{criterion_group, Criterion};
+use criterion::{Criterion, criterion_group};
 
 mod ed25519_benches {
     use super::*;
@@ -15,10 +15,10 @@ mod ed25519_benches {
     use ed25519_dalek::Signer;
     use ed25519_dalek::SigningKey;
     use rand::prelude::ThreadRng;
-    use rand::thread_rng;
+    use rand::rng;
 
     fn sign(c: &mut Criterion) {
-        let mut csprng: ThreadRng = thread_rng();
+        let mut csprng: ThreadRng = rng();
         let keypair: SigningKey = SigningKey::generate(&mut csprng);
         let msg: &[u8] = b"";
 
@@ -26,7 +26,7 @@ mod ed25519_benches {
     }
 
     fn verify(c: &mut Criterion) {
-        let mut csprng: ThreadRng = thread_rng();
+        let mut csprng: ThreadRng = rng();
         let keypair: SigningKey = SigningKey::generate(&mut csprng);
         let msg: &[u8] = b"";
         let sig: Signature = keypair.sign(msg);
@@ -37,7 +37,7 @@ mod ed25519_benches {
     }
 
     fn verify_strict(c: &mut Criterion) {
-        let mut csprng: ThreadRng = thread_rng();
+        let mut csprng: ThreadRng = rng();
         let keypair: SigningKey = SigningKey::generate(&mut csprng);
         let msg: &[u8] = b"";
         let sig: Signature = keypair.sign(msg);
@@ -58,7 +58,7 @@ mod ed25519_benches {
         for size in BATCH_SIZES {
             let name = format!("size={size}");
             group.bench_function(name, |b| {
-                let mut csprng: ThreadRng = thread_rng();
+                let mut csprng: ThreadRng = rng();
                 let keypairs: Vec<SigningKey> = (0..size)
                     .map(|_| SigningKey::generate(&mut csprng))
                     .collect();
@@ -78,7 +78,7 @@ mod ed25519_benches {
     fn verify_batch_signatures(_: &mut Criterion) {}
 
     fn key_generation(c: &mut Criterion) {
-        let mut csprng: ThreadRng = thread_rng();
+        let mut csprng: ThreadRng = rng();
 
         c.bench_function("Ed25519 keypair generation", move |b| {
             b.iter(|| SigningKey::generate(&mut csprng))