+++ x25519-dalek/src/x25519.rs
@@ -17,7 +17,8 @@
 use curve25519_dalek::{edwards::EdwardsPoint, montgomery::MontgomeryPoint, traits::IsIdentity};
 
 use rand_core::CryptoRng;
-use rand_core::RngCore;
+#[cfg(feature = "os_rng")]
+use rand_core::TryRngCore;
 
 #[cfg(feature = "zeroize")]
 use zeroize::{Zeroize, ZeroizeOnDrop};
@@ -29,7 +30,6 @@ use zeroize::{Zeroize, ZeroizeOnDrop};
 /// (in this crate) does *not* automatically happen, but either must be derived
 /// for Drop or explicitly called.
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
-#[cfg_attr(feature = "zeroize", derive(Zeroize))]
 #[derive(PartialEq, Eq, Hash, Copy, Clone, Debug)]
 pub struct PublicKey(pub(crate) MontgomeryPoint);
 
@@ -62,6 +62,13 @@ impl AsRef<[u8]> for PublicKey {
     }
 }
 
+#[cfg(feature = "zeroize")]
+impl Zeroize for PublicKey {
+    fn zeroize(&mut self) {
+        self.0.zeroize();
+    }
+}
+
 /// A short-lived Diffie-Hellman secret key that can only be used to compute a single
 /// [`SharedSecret`].
 ///
@@ -70,7 +77,6 @@ impl AsRef<[u8]> for PublicKey {
 /// are no serialization methods defined.  This means that [`EphemeralSecret`]s can only be
 /// generated from fresh randomness where the compiler statically checks that the resulting
 /// secret is used at most once.
-#[cfg_attr(feature = "zeroize", derive(Zeroize, ZeroizeOnDrop))]
 pub struct EphemeralSecret(pub(crate) [u8; 32]);
 
 impl EphemeralSecret {
@@ -81,16 +87,7 @@ impl EphemeralSecret {
     }
 
     /// Generate a new [`EphemeralSecret`] with the supplied RNG.
-    #[deprecated(
-        since = "2.0.0",
-        note = "Renamed to `random_from_rng`. This will be removed in 2.1.0"
-    )]
-    pub fn new<T: RngCore + CryptoRng>(mut csprng: T) -> Self {
-        Self::random_from_rng(&mut csprng)
-    }
-
-    /// Generate a new [`EphemeralSecret`] with the supplied RNG.
-    pub fn random_from_rng<T: RngCore + CryptoRng>(mut csprng: T) -> Self {
+    pub fn random_from_rng<R: CryptoRng + ?Sized>(csprng: &mut R) -> Self {
         // The secret key is random bytes. Clamping is done later.
         let mut bytes = [0u8; 32];
         csprng.fill_bytes(&mut bytes);
@@ -98,9 +95,9 @@ impl EphemeralSecret {
     }
 
     /// Generate a new [`EphemeralSecret`].
-    #[cfg(feature = "getrandom")]
+    #[cfg(feature = "os_rng")]
     pub fn random() -> Self {
-        Self::random_from_rng(rand_core::OsRng)
+        Self::random_from_rng(&mut rand_core::OsRng.unwrap_err())
     }
 }
 
@@ -110,6 +107,16 @@ impl<'a> From<&'a EphemeralSecret> for PublicKey {
         PublicKey(EdwardsPoint::mul_base_clamped(secret.0).to_montgomery())
     }
 }
+
+impl Drop for EphemeralSecret {
+    fn drop(&mut self) {
+        #[cfg(feature = "zeroize")]
+        self.0.zeroize();
+    }
+}
+
+#[cfg(feature = "zeroize")]
+impl ZeroizeOnDrop for EphemeralSecret {}
 
 /// A Diffie-Hellman secret key which may be used more than once, but is
 /// purposefully not serialiseable in order to discourage key-reuse.  This is
@@ -130,7 +137,6 @@ impl<'a> From<&'a EphemeralSecret> for PublicKey {
 /// secret keys are never reused, which can have very serious security
 /// implications for many protocols.
 #[cfg(feature = "reusable_secrets")]
-#[cfg_attr(feature = "zeroize", derive(Zeroize, ZeroizeOnDrop))]
 #[derive(Clone)]
 pub struct ReusableSecret(pub(crate) [u8; 32]);
 
@@ -143,16 +149,7 @@ impl ReusableSecret {
     }
 
     /// Generate a new [`ReusableSecret`] with the supplied RNG.
-    #[deprecated(
-        since = "2.0.0",
-        note = "Renamed to `random_from_rng`. This will be removed in 2.1.0."
-    )]
-    pub fn new<T: RngCore + CryptoRng>(mut csprng: T) -> Self {
-        Self::random_from_rng(&mut csprng)
-    }
-
-    /// Generate a new [`ReusableSecret`] with the supplied RNG.
-    pub fn random_from_rng<T: RngCore + CryptoRng>(mut csprng: T) -> Self {
+    pub fn random_from_rng<R: CryptoRng + ?Sized>(csprng: &mut R) -> Self {
         // The secret key is random bytes. Clamping is done later.
         let mut bytes = [0u8; 32];
         csprng.fill_bytes(&mut bytes);
@@ -160,9 +157,9 @@ impl ReusableSecret {
     }
 
     /// Generate a new [`ReusableSecret`].
-    #[cfg(feature = "getrandom")]
+    #[cfg(feature = "os_rng")]
     pub fn random() -> Self {
-        Self::random_from_rng(rand_core::OsRng)
+        Self::random_from_rng(&mut rand_core::OsRng.unwrap_mut())
     }
 }
 
@@ -173,6 +170,17 @@ impl<'a> From<&'a ReusableSecret> for PublicKey {
         PublicKey(EdwardsPoint::mul_base_clamped(secret.0).to_montgomery())
     }
 }
+
+#[cfg(feature = "reusable_secrets")]
+impl Drop for ReusableSecret {
+    fn drop(&mut self) {
+        #[cfg(feature = "zeroize")]
+        self.0.zeroize();
+    }
+}
+
+#[cfg(all(feature = "reusable_secrets", feature = "zeroize"))]
+impl ZeroizeOnDrop for ReusableSecret {}
 
 /// A Diffie-Hellman secret key that can be used to compute multiple [`SharedSecret`]s.
 ///
@@ -190,7 +198,6 @@ impl<'a> From<&'a ReusableSecret> for PublicKey {
 /// implications for many protocols.
 #[cfg(feature = "static_secrets")]
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
-#[cfg_attr(feature = "zeroize", derive(Zeroize, ZeroizeOnDrop))]
 #[derive(Clone)]
 pub struct StaticSecret([u8; 32]);
 
@@ -203,16 +210,7 @@ impl StaticSecret {
     }
 
     /// Generate a new [`StaticSecret`] with the supplied RNG.
-    #[deprecated(
-        since = "2.0.0",
-        note = "Renamed to `random_from_rng`. This will be removed in 2.1.0"
-    )]
-    pub fn new<T: RngCore + CryptoRng>(mut csprng: T) -> Self {
-        Self::random_from_rng(&mut csprng)
-    }
-
-    /// Generate a new [`StaticSecret`] with the supplied RNG.
-    pub fn random_from_rng<T: RngCore + CryptoRng>(mut csprng: T) -> Self {
+    pub fn random_from_rng<R: CryptoRng + ?Sized>(csprng: &mut R) -> Self {
         // The secret key is random bytes. Clamping is done later.
         let mut bytes = [0u8; 32];
         csprng.fill_bytes(&mut bytes);
@@ -220,9 +218,9 @@ impl StaticSecret {
     }
 
     /// Generate a new [`StaticSecret`].
-    #[cfg(feature = "getrandom")]
+    #[cfg(feature = "os_rng")]
     pub fn random() -> Self {
-        Self::random_from_rng(rand_core::OsRng)
+        Self::random_from_rng(&mut rand_core::OsRng.unwrap_mut())
     }
 
     /// Extract this key's bytes for serialization.
@@ -263,11 +261,21 @@ impl AsRef<[u8]> for StaticSecret {
     }
 }
 
+#[cfg(feature = "static_secrets")]
+impl Drop for StaticSecret {
+    fn drop(&mut self) {
+        #[cfg(feature = "zeroize")]
+        self.0.zeroize();
+    }
+}
+
+#[cfg(all(feature = "static_secrets", feature = "zeroize"))]
+impl ZeroizeOnDrop for StaticSecret {}
+
 /// The result of a Diffie-Hellman key exchange.
 ///
 /// Each party computes this using their [`EphemeralSecret`] or [`StaticSecret`] and their
 /// counterparty's [`PublicKey`].
-#[cfg_attr(feature = "zeroize", derive(Zeroize, ZeroizeOnDrop))]
 pub struct SharedSecret(pub(crate) MontgomeryPoint);
 
 impl SharedSecret {
@@ -330,6 +338,16 @@ impl AsRef<[u8]> for SharedSecret {
     }
 }
 
+impl Drop for SharedSecret {
+    fn drop(&mut self) {
+        #[cfg(feature = "zeroize")]
+        self.0.zeroize();
+    }
+}
+
+#[cfg(feature = "zeroize")]
+impl ZeroizeOnDrop for SharedSecret {}
+
 /// The bare, byte-oriented x25519 function, exactly as specified in RFC7748.
 ///
 /// This can be used with [`X25519_BASEPOINT_BYTES`] for people who
@@ -340,17 +358,20 @@ impl AsRef<[u8]> for SharedSecret {
 #[cfg_attr(not(feature = "static_secrets"), doc = "```ignore")]
 /// use rand_core::OsRng;
 /// use rand_core::RngCore;
+/// use rand_core::TryRngCore;
 ///
 /// use x25519_dalek::x25519;
 /// use x25519_dalek::StaticSecret;
 /// use x25519_dalek::PublicKey;
 ///
+/// let mut rng = OsRng.unwrap_err();
+///
 /// // Generate Alice's key pair.
-/// let alice_secret = StaticSecret::random_from_rng(&mut OsRng);
+/// let alice_secret = StaticSecret::random_from_rng(&mut rng);
 /// let alice_public = PublicKey::from(&alice_secret);
 ///
 /// // Generate Bob's key pair.
-/// let bob_secret = StaticSecret::random_from_rng(&mut OsRng);
+/// let bob_secret = StaticSecret::random_from_rng(&mut rng);
 /// let bob_public = PublicKey::from(&bob_secret);
 ///
 /// // Alice and Bob should now exchange their public keys.