+++ x25519-dalek/tests/x25519_tests.rs
@@ -181,34 +181,28 @@ fn rfc7748_ladder_test2() {
 mod rand_core {
 
     use super::*;
-    use ::rand_core::OsRng;
+    use ::rand_core::{OsRng, TryRngCore};
 
     #[test]
     fn ephemeral_from_rng() {
-        #[allow(deprecated)]
-        EphemeralSecret::new(OsRng);
-        EphemeralSecret::random_from_rng(OsRng);
+        EphemeralSecret::random_from_rng(&mut OsRng.unwrap_err());
     }
 
     #[test]
     #[cfg(feature = "reusable_secrets")]
     fn reusable_from_rng() {
-        #[allow(deprecated)]
-        ReusableSecret::new(OsRng);
-        ReusableSecret::random_from_rng(OsRng);
+        ReusableSecret::random_from_rng(&mut OsRng.unwrap_err());
     }
 
     #[test]
     #[cfg(feature = "static_secrets")]
     fn static_from_rng() {
-        #[allow(deprecated)]
-        StaticSecret::new(OsRng);
-        StaticSecret::random_from_rng(OsRng);
+        StaticSecret::random_from_rng(&mut OsRng.unwrap_err());
     }
 }
 
-#[cfg(feature = "getrandom")]
-mod getrandom {
+#[cfg(feature = "os_rng")]
+mod os_rng {
 
     use super::*;