+++ ed25519-dalek/CHANGELOG.md
@@ -6,9 +6,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Entries are listed in reverse chronological order per undeprecated major series.
 
-# Unreleased
+# 3.x series
+
+## 3.0.0-pre.0
+
+* Update edition to 2024
+* Update the MSRV from 1.60 to 1.85
+* Update `ed25519` and `signature` deps
+* Remove `std` feature
+* Make signing and verifying keys use `pkcs8::spki::SignatureAlgorithmIdentifier` instead of `DynSignatureAlgorithmIdentifier`
 
 # 2.x series
+
+## 2.2.0
+
+* Add `hazmat`-gated methods `SigningKey::verify_stream()` and `VerifyingKey::verify_stream()`
+* Add `Debug` and `Eq` traits for `hazmat::ExpandedSecretKey`
 
 ## 2.1.1