+++ x25519-dalek/CHANGELOG.md
@@ -2,9 +2,19 @@
 
 Entries are listed in reverse chronological order.
 
+# 3.x Series
+
+## 3.0.0-pre.0
+
+* Update edition to 2024
+* Update the MSRV from 1.60 to 1.85
+* Update `rand_core` dep
+* Remove `Zeroize` impl for `x25519::{EphemeralSecret, ReusableSecret, SharedSecret, StaticSecret}` to prevent misuse. These are now only zeroized on drop.
+* Remove deprecated functions `{Ephemeral,Reusable,Static}Secret::new()`
+
 # 2.x Series
 
-* Note: All `x255919-dalek` 2.x releases are in sync with the underlying `curve25519-dalek` 4.x releases. 
+* Note: All `x255919-dalek` 2.x releases are in sync with the underlying `curve25519-dalek` 4.x releases.
 
 ## 2.0.1
 
@@ -98,5 +108,4 @@ Entries are listed in reverse chronological order.
 
 * Adds support for static and ephemeral keys.
 
-[curve25519-dalek backend]: https://github.com/dalek-cryptography/curve25519-dalek/#backends 
-
+[curve25519-dalek backend]: https://github.com/dalek-cryptography/curve25519-dalek/#backends